requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-docx==0.8.11
openpyxl==3.1.2
PyPDF2==3.0.1
requests==2.31.0
python-dotenv==1.0.0 